Move election types to own file
